+++ pkg/providers/iam/testdata/TestResCIDRBlock/create.tf
@@ -1,0 +1,9 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_cidr_block" "test" {
+  cidr_block = "128.5.6.5/24"
+  comments   = "test"
+  enabled    = true
+}