+++ pkg/providers/iam/testdata/TestResCIDRBlock/importable_without_comments.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_cidr_block" "test" {
+  cidr_block = "128.5.6.6/24"
+  enabled    = true
+}