+++ pkg/providers/iam/testdata/TestResCIDRBlock/missing_cidr_block.tf
@@ -1,3 +1,7 @@
 provider "akamai" {
   edgerc = "../../common/testutils/edgerc"
 }
+
+resource "akamai_iam_cidr_block" "test" {
+  enabled = true
+}